cleanslate/requirements.txt
django-markdownx==2.0.21
